requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.3
click==6.7
gcloud==0.17.0
googleapis-common-protos==1.5.2
html5lib==0.999999999
httplib2==0.10.3
idna==2.5
jws==0.1.3
nltk==3.2.4
oauth2client==3.0.0
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dotenv==0.6.4
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
selenium==3.4.3
six==1.10.0
urllib3==1.21.1
webencodings==0.5.1
